This change applies for tables that take one value for the primary param, but the script is feeding in multiple values for that param.
Consider the URL for the table:

    http://foo.com?param={param}

used with a select

    select * from foo where param in ('1', '2', '3')

In this case, we make three requests to the URL, and merge the results field
by field.

Imagine that the response for each request looks like this

    {'t': 't1',
      'status': 'ok',
      'result': {
        'id': '1',
        'b': 'b1'
      }
    }

By default, we merge the three responses field by field resulting in

    {'t': ['t1','t2','t3'],
      'status': ['ok','ok','ok'],
      'result': [{
        'id': '1',
        'b': 'b1'
      },
      {
        'id': '2',
        'b': 'b2'
      },
      {
        'id': '3',
        'b': 'b3'
      }
    }

This makes sense for some APIs. Consider a different style of response.

    {
      'id': '1',
      'b': 'b1'
    }

In this case, merging the results into an array is more useful.

    [{
        'id': '1',
        'b': 'b1'
      },
      {
        'id': '2',
        'b': 'b2'
      },
      {
        'id': '3',
        'b': 'b3'
      }
    ]

The table can indicate which behavior it wants by prepending '#'.

    http://foo.com?param={#param}

This change implements this variant. The example on http://ql.io also depends on
this fix.