Remove duplicates from in clause.
Here is some background on why this came up now. Recently I fixed joins to include duplicate rows found
which is the correct behavior for joins.

However, prior to this change, a script was using a multi-valued token for a resource with several duplicates
in it. This resulted in duplicate (unintended) results in the join set.

Removing the duplicates from the in clause in the first avoids this. Moreover, duplicate removal from the
in set also makes the request more efficient.